all warnings gone

+++ src/test/unit/unit_load_factor.cpp
@@ -11,7 +11,7 @@ TEST_CASE_TEMPLATE("load_factor", Map, robin_hood::unordered_flat_map<int, int>,
                    robin_hood::unordered_node_map<int, int>) {
 
     Map m;
-    REQUIRE(m.load_factor() == doctest::Approx(0.0));
+    REQUIRE(static_cast<double>(m.load_factor()) == doctest::Approx(0.0));
     for (int i = 0; i < 10000; ++i) {
         m.emplace(i, i);
         REQUIRE(m.load_factor() > 0.0f);
